refactor(tests): split loreStore tests into focused modules (#818)
Split loreStore.test.ts (518 lines) into:
- basic.test.ts (75 lines) - Add, get by category, clear operations
- crud.test.ts (107 lines) - Update, delete, AI context generation
- search.test.ts (128 lines) - Duplicate detection and search functionality
- advanced.test.ts (254 lines) - Import/export, validation, history, structured lore
- testHelpers.ts (85 lines) - Reusable test utilities and data factories

Also created:
- endingGenerator.testHelpers.ts (140 lines) - Mock data factories for ending tests

Benefits:
- All test files now under 300 lines (max is 254)
- Improved organization by feature area
- Reduced duplication through shared helpers
- Easier to locate specific test scenarios

Progress: 4 of 48 files refactored (8%)
